try conversion to 64 bit ints

+++ cvxpy/interface/cvxopt_interface/sparse_matrix_interface.py
@@ -48,8 +48,8 @@ class SparseMatrixInterface(DenseMatrixInterface):
         elif sp.issparse(value):
             value = value.tocoo()
             V = cvxopt.matrix(value.data.astype('float64'), tc='d')
-            I = cvxopt.matrix(value.row.astype('int32'), tc='i')
-            J = cvxopt.matrix(value.col.astype('int32'), tc='i')
+            I = cvxopt.matrix(value.row.astype('int64'), tc='i')
+            J = cvxopt.matrix(value.col.astype('int64'), tc='i')
             return cvxopt.spmatrix(V, I, J, value.shape, tc='d')
         else: # Lists.
             return cvxopt.sparse(value, tc='d')
